% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{vegseg_config}
\alias{vegseg_config}
\title{Default config for vegseg (override values as needed)}
\usage{
vegseg_config(
  las_path = "data/trees.las",
  out_dir = "data/ds_hag4",
  out_pred_dir = "data/output_predictions",
  model_path = "data/model/best_model.pth",
  device = NULL,
  block_size = 6,
  stride = 1,
  sample_n = 4096,
  repeat_per_tile = 4,
  min_pts_tile = 512,
  val_split = 0.15,
  test_split = 0.1,
  seed = 42,
  batch_size = 16,
  epochs = 20,
  learning_rate = 1e-05,
  weight_decay = 1e-04,
  cell_size = 0.25,
  quantile = 0.05
)
}
\description{
Default config for vegseg (override values as needed)
}
