% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.r
\name{vegseg_py_setup}
\alias{vegseg_py_setup}
\title{Create/use a venv and install Python deps listed in inst/python/requirements.txt}
\usage{
vegseg_py_setup(
  envdir = file.path(system.file("python", package = "vegseg"), ".venv")
)
}
\description{
Create/use a venv and install Python deps listed in inst/python/requirements.txt
}
