% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{vegseg_train}
\alias{vegseg_train}
\title{Train the vegseg model (auto-preprocess if NPZ tiles are missing)}
\usage{
vegseg_train(cfg, setup_env = FALSE)
}
\arguments{
\item{cfg}{list from vegseg_config()}

\item{setup_env}{logical; set TRUE to create/use the package venv}
}
\value{
list(best_oa, best_epoch, ckpt_path)
}
\description{
Train the vegseg model (auto-preprocess if NPZ tiles are missing)
}
