% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{vegseg_predict}
\alias{vegseg_predict}
\title{Predict on a LAS and write a predicted LAS}
\usage{
vegseg_predict(
  cfg = vegseg_config(),
  new_in_las = NULL,
  new_out_las = NULL,
  mode = "overwrite",
  setup_env = TRUE
)
}
\arguments{
\item{mode}{"overwrite" (classification) or "extra" (adds pred_label)}
}
\description{
Predict on a LAS and write a predicted LAS
}
