% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{vegseg_train}
\alias{vegseg_train}
\title{Train model (builds NPZ if missing)}
\usage{
vegseg_train(cfg = vegseg_config(), setup_env = TRUE)
}
\arguments{
\item{cfg}{list from vegseg_config()}
}
\description{
Train model (builds NPZ if missing)
}
